version 1.3

+++ src/Context/PointsContext.tsx
@@ -15,6 +15,7 @@ const PointsContext = createContext<PointsContextType | undefined>(undefined);
 
 export const usePointsContext = () => {
   const context = useContext(PointsContext);
+  // console.log(context); // Adicione isto para verificar o valor
   if (!context) {
     throw new Error('usePointsContext must be used within a PointsProvider');
   }
@@ -36,35 +37,25 @@ export const PointsProvider = ({ children }: PointsProviderProps) => {
   }, []);
   
 
-  const handleAddPoint = async (productId: number) => {
-    try {
-      const response = await api.post(`/addPoint/${productId}`);
-      if (response.status === 201) {
-        setPointsAdded((prevPoints) => {
-          const currentTime = Date.now();
-          const updatedPoints = prevPoints[productId] || [];
-          const filteredPoints = updatedPoints.filter(
-            (timestamp) => currentTime - timestamp <= 5 * 60 * 60 * 1000 // 5 Horas
-          );
-
-          if (filteredPoints.length < 4) {
-            filteredPoints.push(currentTime);
-            const newPointsAdded = {
-              ...prevPoints,
-              [productId]: filteredPoints,
-            };
-            localStorage.setItem('pointsAdded', JSON.stringify(newPointsAdded));
-            return newPointsAdded;
-          }
-          return prevPoints;
-        });
+  const handleAddPoint = (productId: number) => {
+    const currentTime = Date.now();
+    setPointsAdded((prevPoints) => {
+      const updatedPoints = prevPoints[productId]?.filter(
+        timestamp => currentTime - timestamp <= 5 * 60 * 60 * 1000
+      ) || [];
+  
+      if (updatedPoints.length < 4) {
+        const newPoints = { ...prevPoints, [productId]: [...updatedPoints, currentTime] };
+        localStorage.setItem('pointsAdded', JSON.stringify(newPoints)); // Sincronizando com localStorage
+        return newPoints;
       }
-    } catch (error) {
-      console.error('Erro ao adicionar ponto:', error);
-    }
+      return prevPoints;
+    });
   };
+  
 
   const handleRemovePoint = async (productId: number) => {
+    if (!pointsAdded.hasOwnProperty(productId)) return;
     try {
       const response = await api.delete(`/deletePoint/${productId}`);
       if (response.status === 200) {
@@ -72,7 +63,7 @@ export const PointsProvider = ({ children }: PointsProviderProps) => {
           const currentTime = Date.now();
           const updatedPoints = prevPoints[productId] || [];
           const filteredPoints = updatedPoints.filter(
-            (timestamp) => currentTime - timestamp <= 60 * 1000
+            (timestamp) => currentTime - timestamp <= 5 * 60 * 60 * 1000 // 5 Horas
           );
 
           if (filteredPoints.length > 0) {
